+++ lib/Election.php
@@ -34,16 +34,16 @@ class Election
 /////////// STATICS METHODS ///////////
 
     // Change max parse iteration
-    public static function setMaxParseIteration (?int $value) : ?int
+    public static function setMaxParseIteration (?int $maxParseIterations) : ?int
     {
-        self::$_maxParseIteration = $value;
+        self::$_maxParseIteration = $maxParseIterations;
         return self::$_maxParseIteration;
     }
 
     // Change max vote number
-    public static function setMaxVoteNumber (?int $value) : ?int
+    public static function setMaxVoteNumber (?int $maxVotesNumber) : ?int
     {
-        self::$_maxVoteNumber = $value;
+        self::$_maxVoteNumber = $maxVotesNumber;
         return self::$_maxVoteNumber;
     }
 
@@ -82,9 +82,9 @@ class Election
             '_Candidates' => $this->_Candidates,
             '_Votes' => $this->_Votes,
 
-            '_i_CandidateId' => $this->_i_CandidateId,
             '_State' => $this->_State,
             '_objectVersion' => $this->_objectVersion,
+            '_AutomaticNewCandidateName' => $this->_AutomaticNewCandidateName,
 
             '_ImplicitRanking' => $this->_ImplicitRanking,
             '_VoteWeightRule' => $this->_VoteWeightRule,
@@ -108,7 +108,7 @@ class Election
         $this->_Candidates = $data['_Candidates'];
         $this->_Votes = $data['_Votes'];
 
-        $this->_i_CandidateId = $data['_i_CandidateId'];
+        $this->_AutomaticNewCandidateName = $data['_AutomaticNewCandidateName'];
         $this->_State = $data['_State'];
         $this->_objectVersion = $data['_objectVersion'];
 
@@ -234,13 +234,13 @@ class Election
 
     /////////// VOTE CONSTRAINT ///////////
 
-    public function addConstraint (string $class) : bool
+    public function addConstraint (string $constraintClass) : bool
     {
-        if ( !class_exists($class) ) :
+        if ( !class_exists($constraintClass) ) :
             throw new CondorcetException(27);
-        elseif ( !is_subclass_of($class, __NAMESPACE__.'\\VoteConstraint') ) :
+        elseif ( !is_subclass_of($constraintClass, __NAMESPACE__.'\\VoteConstraint') ) :
             throw new CondorcetException(28);
-        elseif (in_array($class,$this->getConstraints(), true)) :
+        elseif (in_array($constraintClass,$this->getConstraints(), true)) :
             throw new CondorcetException(29);
         endif;
 
@@ -248,7 +248,7 @@ class Election
             $this->cleanupCompute();;
         endif;
 
-        $this->_Constraints[] = $class;
+        $this->_Constraints[] = $constraintClass;
 
         return true;
     }